[ADD] Configure products with variants
The module configures product variants based on attributes specified by the customer.
We need to add an attribute to identify every part of the product with an ID and create a code based on all these parts.
This commit is not complete because it does not generate the code for all the variants yet, but the attribute is already added.