[ADD] Module to create job number and plant code
This module creates a job number for every sales order. This field is composed of a prefix that the user chooses in a selection field;
a sequence that starts at 13500 and it automatically counts up; and a sufix that the user chooses in a selection field.
The user must be able to add values for both the prefix and the sufix using Studio.
The sequence is created using the ir.sequence model.

For the plant code, the field is added to the customer (res.partner) and is composed of two parts:
The first one is a prefix that consists of the first three non-special characters of the companys name. (Example: AT&T would have a prefix ATT).
The second part of the code is a sequence with a padding value of 5 and using this nomenclature: XXX-XX starting in 001-01.
An example of the complete code would be ATT001-01 and if there is another company that would have the prefix ATT, its code would be:
ATT001-02. This plant code is computed once the first Sale Order is confirmed for that customer.

The modules used in this task were sale_order and res_partner.

Both res_partner and sale order views were inherited to add the fields to those views.