[IMP] product_specific_lots: An auto-product specific lot setter
When manufacturing products, the finished goods need to have lot numbers that correlate with the product for record keeping purposes. Additionally due to the volume and daily operations, it needs to be automatically assignable without knowing what the previous lot number for that item was. They would then count up independent of each other.