[IMP] Corrected some PR comments and changed the logic of Plant Code.
I used to have a computed field in the res_partner model that indicated if the customer had confirmed their first SO.
Now this field is not computed and it is modified in the sale_order model once the quotation is confirmed.

In the sale_order model I overwrote the create method in order to add a field when created but now that method is called using super and then the field is stored.