[ADD] Added module to create automated product name and UPC.
This module creates the product name based on their category and each category has a sequence so the product name is the category + the sequence value.
The UPC is the barcode field and it is created like the product name but instead of using the category, a new gender field is used.