[IMP] New York P&W: Automated product name & UPC
The customer generates product names when he orders new products from the suppliers.
The product name is a code and is linked to the product category (names must be unique).
The customer also needs to generate UPC automatically. He generates UPC and then sends it back to its manufacturer who will print it on the labels.