[IMP] code_survey: safe_compile module for safe(r) untrusted code exec
This commit contains the code that handles the untrusted code execution from the user in the extended survey module.

Basically, the code relies on the existing safe_eval module, but adds some necessary features (such as importing modules and classes) for code evaluation and testing.
The class that wraps the candidate's solution is built at runtime, and cherry-picked from the code object provided by the compile() function

Also added some code formatting corrections.